Move to using raw USB device lookup
gousb would be useful if we could use it inside a container, but to do so would mean linking libusb - something we desperately want to avoid inside a container due to the substantial increase in size.

This solution just enumerates the /sys/bus. Linux-only.
